refactor: `ResourceSettings` interface into `settings.ts`
- Moved the `ResourceSettings` interface from it's separate `resource-settings.ts` file into the `settings.ts` file to consolidate all settings in file.

- Added docblocks to all settings.

- Removed the `resource-settings.ts` file.

- Updated all `ResourceSettings` imports to the appropriate file path.
