--- src/interfaces/resource-settings.ts
@@ -1,14 +1,0 @@
-export interface ResourceSettings {
-  workspaceRoot: string | null;
-  fixerEnable: boolean;
-  fixerArguments: string[];
-  snifferEnable: boolean;
-  snifferArguments: string[];
-  executablePathCBF: string;
-  executablePathCS: string;
-  composerJsonPath: string;
-  standard: string | null;
-  autoRulesetSearch: boolean;
-  allowedAutoRulesets: string[];
-  excludeGlobs: string[];
-}
