+++ src/resolvers/standards-path-resolver.ts
@@ -2,7 +2,7 @@ import fs from 'node:fs/promises';
 import { TextDocument, window, workspace } from 'vscode';
 import { ConsoleError } from '../interfaces/console-error';
 import { PathResolver } from '../interfaces/path-resolver';
-import { ResourceSettings } from '../interfaces/resource-settings';
+import { ResourceSettings } from '../interfaces/settings';
 import { logger } from '../logger';
 import { isSingleFileMode } from '../settings';
 import { getErrorCodeDescription } from '../utils/error-handling/error-helpers';
